+++ Firestore/Example/Tests/Integration/API/FIRDatabaseTests.mm
@@ -1224,13 +1224,13 @@ using firebase::firestore::util::TimerId;
   [self awaitExpectations];
 }
 
-- (void)testClientCallsAfterShutdownFail {
+- (void)testClientCallsAfterTerminationFail {
   FIRDocumentReference *doc = [self documentRef];
   FIRFirestore *firestore = doc.firestore;
 
   [firestore enableNetworkWithCompletion:[self completionForExpectationWithName:@"Enable network"]];
   [self awaitExpectations];
-  [firestore shutdownWithCompletion:[self completionForExpectationWithName:@"Shutdown"]];
+  [firestore terminateWithCompletion:[self completionForExpectationWithName:@"Terminate"]];
   [self awaitExpectations];
 
   XCTAssertThrowsSpecific(
@@ -1238,7 +1238,7 @@ using firebase::firestore::util::TimerId;
         [firestore disableNetworkWithCompletion:^(NSError *error){
         }];
       },
-      NSException, @"The client has already been shutdown.");
+      NSException, @"The client has already been terminated.");
 }
 
 - (void)testMaintainsPersistenceAfterRestarting {
@@ -1251,9 +1251,9 @@ using firebase::firestore::util::TimerId;
   NSDictionary<NSString *, id> *initialData = @{@"foo" : @"42"};
   [self writeDocumentRef:doc data:initialData];
 
-  // -clearPersistence() requires Firestore to be shut down. Shutdown FIRApp and remove the
+  // -clearPersistence() requires Firestore to be terminated. Shutdown FIRApp and remove the
   // firestore instance to emulate the way an end user would do this.
-  [self shutdownFirestore:firestore];
+  [self terminateFirestore:firestore];
   [self.firestores removeObject:firestore];
   [self deleteApp:app];
 
@@ -1278,9 +1278,9 @@ using firebase::firestore::util::TimerId;
   NSDictionary<NSString *, id> *initialData = @{@"foo" : @"42"};
   [self writeDocumentRef:doc data:initialData];
 
-  // -clearPersistence() requires Firestore to be shut down. Shutdown FIRApp and remove the
+  // -clearPersistence() requires Firestore to be terminated. Shutdown FIRApp and remove the
   // firestore instance to emulate the way an end user would do this.
-  [self shutdownFirestore:firestore];
+  [self terminateFirestore:firestore];
   [self.firestores removeObject:firestore];
   [firestore
       clearPersistenceWithCompletion:[self completionForExpectationWithName:@"Enable network"]];
@@ -1332,7 +1332,7 @@ using firebase::firestore::util::TimerId;
       @{@"owner" : @{@"name" : @"Jonny", @"email" : @"abc@xyz.com"}};
   [self writeDocumentRef:[firestore documentWithPath:@"abc/123"] data:data];
 
-  [self shutdownFirestore:firestore];
+  [self terminateFirestore:firestore];
 
   // Create a new instance, check it's a different instance.
   FIRFirestore *newInstance = [FIRFirestore firestoreForApp:app];
@@ -1345,7 +1345,7 @@ using firebase::firestore::util::TimerId;
   XCTAssertTrue([data isEqualToDictionary:[snapshot data]]);
 }
 
-- (void)testAppDeleteLeadsToFirestoreShutdown {
+- (void)testAppDeleteLeadsToFirestoreTermination {
   FIRApp *app = testutil::AppForUnitTesting(util::MakeString([FSTIntegrationTestCase projectID]));
   FIRFirestore *firestore = [FIRFirestore firestoreForApp:app];
   firestore.settings = [FSTIntegrationTestCase settings];
@@ -1355,33 +1355,33 @@ using firebase::firestore::util::TimerId;
 
   [self deleteApp:app];
 
-  XCTAssertTrue(firestore.wrapped->client()->is_shutdown());
+  XCTAssertTrue(firestore.wrapped->client()->is_terminated());
 }
 
-- (void)testShutdownCanBeCalledMultipleTimes {
+- (void)testTerminateCanBeCalledMultipleTimes {
   FIRApp *app = testutil::AppForUnitTesting(util::MakeString([FSTIntegrationTestCase projectID]));
   FIRFirestore *firestore = [FIRFirestore firestoreForApp:app];
 
-  [firestore shutdownWithCompletion:[self completionForExpectationWithName:@"Shutdown1"]];
+  [firestore terminateWithCompletion:[self completionForExpectationWithName:@"Terminate1"]];
   [self awaitExpectations];
   XCTAssertThrowsSpecific(
       {
         [firestore disableNetworkWithCompletion:^(NSError *error){
         }];
       },
-      NSException, @"The client has already been shutdown.");
+      NSException, @"The client has already been terminated.");
 
-  [firestore shutdownWithCompletion:[self completionForExpectationWithName:@"Shutdown2"]];
+  [firestore terminateWithCompletion:[self completionForExpectationWithName:@"Terminate2"]];
   [self awaitExpectations];
   XCTAssertThrowsSpecific(
       {
         [firestore enableNetworkWithCompletion:^(NSError *error){
         }];
       },
-      NSException, @"The client has already been shutdown.");
+      NSException, @"The client has already been terminated.");
 }
 
-- (void)testCanRemoveListenerAfterShutdown {
+- (void)testCanRemoveListenerAfterTermination {
   FIRApp *app = testutil::AppForUnitTesting(util::MakeString([FSTIntegrationTestCase projectID]));
   FIRFirestore *firestore = [FIRFirestore firestoreForApp:app];
   firestore.settings = [FSTIntegrationTestCase settings];
@@ -1393,7 +1393,7 @@ using firebase::firestore::util::TimerId;
       [doc addSnapshotListener:[accumulator valueEventHandler]];
   [accumulator awaitEventWithName:@"Snapshot"];
 
-  [firestore shutdownWithCompletion:[self completionForExpectationWithName:@"shutdown"]];
+  [firestore terminateWithCompletion:[self completionForExpectationWithName:@"terminate"]];
   [self awaitExpectations];
 
   // This should proceed without error.