+++ Firestore/Source/API/FIRFirestore.mm
@@ -305,17 +305,17 @@ NS_ASSUME_NONNULL_BEGIN
   return (__bridge FIRFirestore *)firestore->extension();
 }
 
-- (void)shutdownInternalWithCompletion:(nullable void (^)(NSError *_Nullable error))completion {
-  _firestore->Shutdown(util::MakeCallback(completion));
+- (void)terminateInternalWithCompletion:(nullable void (^)(NSError *_Nullable error))completion {
+  _firestore->Terminate(util::MakeCallback(completion));
 }
 
-- (void)shutdownWithCompletion:(nullable void (^)(NSError *_Nullable error))completion {
+- (void)terminateWithCompletion:(nullable void (^)(NSError *_Nullable error))completion {
   id<FSTFirestoreInstanceRegistry> strongRegistry = _registry;
   if (strongRegistry) {
     [strongRegistry
         removeInstanceWithDatabase:util::MakeNSString(_firestore->database_id().database_id())];
   }
-  [self shutdownInternalWithCompletion:completion];
+  [self terminateInternalWithCompletion:completion];
 }
 
 @end